Added count unit test, fixed bugs
